requirements.txt
streamlit
pandas
numpy
google-generativeai
requests
faiss-cpu